+++ torchtune/training/__init__.py
@@ -4,8 +4,10 @@
 # This source code is licensed under the BSD-style license found in the
 # LICENSE file in the root directory of this source tree.
 
+from torchtune.training.pooling import get_unmasked_sequence_lengths
 from torchtune.training.quantization import get_quantizer_mode
 
 __all__ = [
     "get_quantizer_mode",
+    "get_unmasked_sequence_lengths",
 ]
