+++ torchtune/utils/__init__.py
@@ -62,7 +62,6 @@ from .memory import (  # noqa
     register_optim_in_bwd_hooks,
     set_activation_checkpointing,
 )
-from .pooling import get_unmasked_sequence_lengths
 
 from .precision import get_dtype, set_default_dtype, validate_expected_param_dtype
 from .seed import set_seed
@@ -80,7 +79,6 @@ __all__ = [
     "is_distributed",
     "lora_fsdp_wrap_policy",
     "get_full_finetune_fsdp_wrap_policy",
-    "get_unmasked_sequence_lengths",
     "set_activation_checkpointing",
     "set_default_dtype",
     "set_seed",
